# feat: implement analytics and performance monitoring with vercel and react scan