--- a/package.json
+++ b/package.json
@@ -6,7 +6,8 @@
 		"dev": "next dev",
 		"build": "next build",
 		"start": "next start",
-		"lint": "next lint"
+		"lint": "next lint",
+		"scan": "next dev -p 3333 & sleep 5 && npx react-scan@latest http://localhost:3333"
 	},
 	"dependencies": {
 		"@hookform/resolvers": "^3.10.0",
@@ -31,6 +32,7 @@
 		"@radix-ui/react-toggle-group": "^1.1.1",
 		"@radix-ui/react-tooltip": "^1.1.6",
 		"@tanstack/react-table": "^8.20.6",
+		"@vercel/analytics": "^1.4.1",
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
 		"cmdk": "1.0.0",