… app/layout.tsx
… pnpm-lock.yaml