Closes #62

Brief overview:
This PR implements comprehensive analytics and performance monitoring using Vercel Analytics for production metrics and React Scan for development-time performance tracking.

### Changes Made
- ⚡️ Add React Scan for real-time performance monitoring in development
- 📊 Integrate Vercel Analytics for production metrics
- ⚠️ Set up error boundary with analytics reporting
- 📈 Configure performance budgets for critical components
- 🔍 Add custom event tracking for key user interactions

### Technical Details
- Implemented React Scan configuration with performance budgets for:
  - MetricGrid (max 5 renders, 16ms budget)
  - ActivityFeed (max 3 renders, 20ms budget)
  - PerformanceChart (max 4 renders, 16ms budget)
  - AlertManagement (max 3 renders, 16ms budget)
- Added custom event tracking for:
  - Contract switching
  - Alert rule management
  - Chart interactions
  - Filter usage
- Set up error boundary with Vercel Analytics integration for error reporting
- Configured development-time performance monitoring with React Scan toolbar

### Testing
- [x] Verify React Scan integration in development environment
- [x] Test error boundary with simulated errors
- [x] Confirm custom event tracking triggers
- [x] Validate performance budget warnings
- [x] Check Vercel Analytics integration
- [x] Test analytics in different environments (dev/prod)

### Performance Impact
- Added minimal runtime overhead with Vercel Analytics
- React Scan only runs in development
- Performance budgets ensure consistent UI responsiveness
- Error boundary provides graceful fallbacks